fix(rc): Exposed admin.remoteConfig namespace
