+++ src/remote-config/remote-config.ts
@@ -17,17 +17,18 @@
 import { FirebaseServiceInterface, FirebaseServiceInternalsInterface } from '../firebase-service';
 import { FirebaseApp } from '../firebase-app';
 import * as validator from '../utils/validator';
-import {
-  RemoteConfigTemplate,
-  RemoteConfigParameter,
-  RemoteConfigCondition,
-  RemoteConfigParameterGroup,
-  ListVersionsOptions,
-  ListVersionsResult,
-  RemoteConfigUser,
-  Version,
-} from './remote-config-api-client';
+import { remoteConfig } from './index';
 import { FirebaseRemoteConfigError, RemoteConfigApiClient } from './remote-config-api-client-internal';
+
+import RemoteConfigTemplate = remoteConfig.RemoteConfigTemplate;
+import RemoteConfigParameter = remoteConfig.RemoteConfigParameter;
+import RemoteConfigCondition = remoteConfig.RemoteConfigCondition;
+import RemoteConfigParameterGroup = remoteConfig.RemoteConfigParameterGroup;
+import ListVersionsOptions = remoteConfig.ListVersionsOptions;
+import ListVersionsResult = remoteConfig.ListVersionsResult;
+import RemoteConfigUser = remoteConfig.RemoteConfigUser;
+import Version = remoteConfig.Version;
+import RemoteConfigInterface = remoteConfig.RemoteConfig;
 
 /**
  * Internals of an RemoteConfig service instance.
@@ -47,13 +48,13 @@ class RemoteConfigInternals implements FirebaseServiceInternalsInterface {
 /**
  * Remote Config service bound to the provided app.
  */
-export class RemoteConfig implements FirebaseServiceInterface {
+export class RemoteConfig implements FirebaseServiceInterface, RemoteConfigInterface {
   public readonly INTERNAL: RemoteConfigInternals = new RemoteConfigInternals();
 
   private readonly client: RemoteConfigApiClient;
 
   /**
-   * @param {FirebaseApp} app The app for this RemoteConfig service.
+   * @param app The app for this RemoteConfig service.
    * @constructor
    */
   constructor(readonly app: FirebaseApp) {
@@ -76,9 +77,9 @@ export class RemoteConfig implements FirebaseServiceInterface {
   /**
    * Gets the requested version of the {@link admin.remoteConfig.RemoteConfigTemplate
     * `RemoteConfigTemplate`} of the project.
-   * 
+   *
    * @param versionNumber Version number of the Remote Config template to look up.
-   * 
+   *
    * @return A promise that fulfills with a `RemoteConfigTemplate`.
    */
   public getTemplateAtVersion(versionNumber: number | string): Promise<RemoteConfigTemplate> {
@@ -126,7 +127,7 @@ export class RemoteConfig implements FirebaseServiceInterface {
    * Rolls back a project's published Remote Config template to the specified version.
    * A rollback is equivalent to getting a previously published Remote Config
    * template and re-publishing it using a force update.
-   * 
+   *
    * @param versionNumber The version number of the Remote Config template to roll back to.
    *    The specified version number must be lower than the current version number, and not have
    *    been deleted due to staleness. Only the last 300 versions are stored.
@@ -142,12 +143,12 @@ export class RemoteConfig implements FirebaseServiceInterface {
   }
 
   /**
-   * Gets a list of Remote Config template versions that have been published, sorted in reverse 
+   * Gets a list of Remote Config template versions that have been published, sorted in reverse
    * chronological order. Only the last 300 versions are stored.
-   * All versions that correspond to non-active Remote Config templates (i.e., all except the 
+   * All versions that correspond to non-active Remote Config templates (i.e., all except the
    * template that is being fetched by clients) are also deleted if they are older than 90 days.
-   * 
-   * @param {ListVersionsOptions} options Optional options object for getting a list of versions.
+   *
+   * @param options Optional options object for getting a list of versions.
    * @return A promise that fulfills with a `ListVersionsResult`.
    */
   public listVersions(options?: ListVersionsOptions): Promise<ListVersionsResult> {
