+++ test/unit/remote-config/remote-config.spec.ts
@@ -22,17 +22,17 @@ import * as sinon from 'sinon';
 import { RemoteConfig } from '../../../src/remote-config/remote-config';
 import { FirebaseApp } from '../../../src/firebase-app';
 import * as mocks from '../../resources/mocks';
-import {
-  RemoteConfigTemplate,
-  RemoteConfigCondition,
-  TagColor,
-  ListVersionsResult,
-} from '../../../src/remote-config/remote-config-api-client';
+import { remoteConfig } from '../../../src/remote-config/index';
 import {
   FirebaseRemoteConfigError,
   RemoteConfigApiClient
 } from '../../../src/remote-config/remote-config-api-client-internal';
 import { deepCopy } from '../../../src/utils/deep-copy';
+
+import RemoteConfigTemplate = remoteConfig.RemoteConfigTemplate;
+import RemoteConfigCondition = remoteConfig.RemoteConfigCondition;
+import TagColor = remoteConfig.TagColor;
+import ListVersionsResult = remoteConfig.ListVersionsResult;
 
 const expect = chai.expect;
 
@@ -82,7 +82,7 @@ describe('RemoteConfig', () => {
       {
         name: 'ios',
         expression: 'device.os == \'ios\'',
-        tagColor: TagColor.BLUE,
+        tagColor: 'BLUE',
       },
     ],
     parameters: {
@@ -102,7 +102,7 @@ describe('RemoteConfig', () => {
     conditions: [{
       name: 'ios',
       expression: 'device.os == \'ios\'',
-      tagColor: TagColor.PINK,
+      tagColor: 'PINK',
     }],
     parameters: {
       // eslint-disable-next-line @typescript-eslint/camelcase
@@ -484,7 +484,7 @@ describe('RemoteConfig', () => {
       expect(newTemplate.conditions.length).to.equal(1);
       expect(newTemplate.conditions[0].name).to.equal('ios');
       expect(newTemplate.conditions[0].expression).to.equal('device.os == \'ios\'');
-      expect(newTemplate.conditions[0].tagColor).to.equal(TagColor.BLUE);
+      expect(newTemplate.conditions[0].tagColor).to.equal('BLUE');
       // verify that the etag is unchanged
       expect(newTemplate.etag).to.equal('etag-123456789012-5');
       // verify that the etag is read-only
@@ -506,7 +506,7 @@ describe('RemoteConfig', () => {
       const cond = c as RemoteConfigCondition;
       expect(cond.name).to.equal('ios');
       expect(cond.expression).to.equal('device.os == \'ios\'');
-      expect(cond.tagColor).to.equal(TagColor.BLUE);
+      expect(cond.tagColor).to.equal('BLUE');
     });
   });
 
@@ -639,7 +639,7 @@ describe('RemoteConfig', () => {
           expect(template.conditions.length).to.equal(1);
           expect(template.conditions[0].name).to.equal('ios');
           expect(template.conditions[0].expression).to.equal('device.os == \'ios\'');
-          expect(template.conditions[0].tagColor).to.equal(TagColor.BLUE);
+          expect(template.conditions[0].tagColor).to.equal('BLUE');
           expect(template.etag).to.equal('etag-123456789012-5');
           // verify that etag is read-only
           expect(() => {
@@ -670,7 +670,7 @@ describe('RemoteConfig', () => {
           const cond = c as RemoteConfigCondition;
           expect(cond.name).to.equal('ios');
           expect(cond.expression).to.equal('device.os == \'ios\'');
-          expect(cond.tagColor).to.equal(TagColor.BLUE);
+          expect(cond.tagColor).to.equal('BLUE');
 
           const parsed = JSON.parse(JSON.stringify(template));
           const expectedTemplate = deepCopy(REMOTE_CONFIG_RESPONSE);
